# backend/openapi.yaml
openapi: 3.0.3
info:
  title: FlexLiving Reviews API
  description: |
    A comprehensive review management system for FlexLiving properties with integrations for Hostaway and Google Reviews.
    
    This API provides endpoints for managing property reviews, statistics, and external integrations.
    
    ## Features
    - Review management with filtering and pagination
    - Property management with review aggregation
    - Hostaway integration for review synchronization
    - Google Places API integration for review exploration
    - Comprehensive statistics and analytics
    
  version: 1.0.0
  contact:
    name: FlexLiving Development Team
    email: dev@flexliving.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5001
    description: Development server
  - url: ""
    description: Production server (to be configured)
...
paths:
...
  /api/reviews:
    get:
      tags:
        - Reviews
      summary: Get all reviews with filtering and pagination
      description: Retrieve reviews with optional filtering by property, channel, rating, status, and date range
      operationId: getAllReviews
      parameters:
        - name: listingId
          in: query
          description: Filter by property listing ID
          required: false
          schema:
            type: string
            example: "PROP001"
        - name: channel
          in: query
          description: Filter by booking channel
          required: false
          schema:
            type: string
            enum: [Airbnb, Booking.com, Direct, Vrbo, Expedia, Google, Other]
            example: "Airbnb"
        - name: rating
          in: query
          description: Filter by minimum rating
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 10
            example: 8
        - name: status
          in: query
          description: Filter by review status
          required: false
          schema:
            type: string
            enum: [published, pending, rejected]
            example: "published"
        - name: showOnWebsite
          in: query
          description: Filter by website visibility
          required: false
          schema:
            type: boolean
            example: true
        - name: startDate
          in: query
          description: Filter reviews from this date (ISO 8601)
          required: false
          schema:
            type: string
            format: date
            example: "2024-01-01"
        - name: endDate
          in: query
          description: Filter reviews until this date (ISO 8601)
          required: false
          schema:
            type: string
            format: date
            example: "2024-12-31"
        - name: sortBy
          in: query
          description: Sort field
          required: false
          schema:
            type: string
            enum: [submittedAt, rating, guestName, channel]
            default: submittedAt
            example: "submittedAt"
        - name: order
          in: query
          description: Sort order
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
            example: "desc"
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            example: 1
        - name: limit
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
            example: 10
      responses:
        '200':
          description: Reviews retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedReviewsResponse'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/properties:
    get:
      tags:
        - Properties
      summary: Get all properties
      description: Retrieve all properties with optional filtering and sorting
      operationId: getAllProperties
      parameters:
        - name: isActive
          in: query
          description: Filter by active status
          required: false
          schema:
            type: boolean
            example: true
        - name: sortBy
          in: query
          description: Sort field
          required: false
          schema:
            type: string
            enum: [name, avgRating, totalReviews, createdAt]
            default: avgRating
            example: "avgRating"
        - name: order
          in: query
          description: Sort order
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
            example: "desc"
      responses:
        '200':
          description: Properties retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertiesSuccessResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/hostaway/reviews:
    get:
      tags:
        - Hostaway
      summary: Fetch reviews from Hostaway
      description: Retrieve reviews from Hostaway API or mock data for testing
      operationId: getHostawayReviews
      parameters:
        - name: useApi
          in: query
          description: Whether to use real Hostaway API or mock data
          required: false
          schema:
            type: boolean
            default: false
            example: false
      responses:
        '200':
          description: Hostaway reviews retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HostawayReviewsResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/google/reviews:
    get:
      tags:
        - Google
      summary: Get Google Reviews for a place
      description: Retrieve Google Reviews for a specific place using Google Places API
      operationId: getGoogleReviews
      parameters:
        - name: placeId
          in: query
          description: Google Place ID
          required: true
          schema:
            type: string
            example: "ChIJN1t_tDeuEmsRUsoyG83frY4"
      responses:
        '200':
          description: Google reviews retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoogleReviewsResponse'
        '400':
          description: Place ID is required or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to fetch Google reviews
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success]
          example: "success"
        message:
          type: string
          example: "Operation completed successfully"
        data:
          type: object
          description: Response data (varies by endpoint)
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          enum: [error]
          example: "error"
        message:
          type: string
          example: "An error occurred"
        error:
          type: string
          description: Additional error details
          example: "Validation failed"
...
    ReviewCategory:
      type: object
      properties:
        category:
          type: string
          enum: [cleanliness, accuracy, check_in, communication, location, value, respect_house_rules]
          example: "cleanliness"
        rating:
          type: integer
          minimum: 1
          maximum: 10
          example: 9
    Review:
      type: object
      properties:
        _id:
          type: string
          example: "60f7b3b3b3b3b3b3b3b3b3b3"
        externalId:
          type: integer
          example: 12345
        type:
          type: string
          enum: [guest-to-host, host-to-guest]
          example: "guest-to-host"
        status:
          type: string
          enum: [published, pending, rejected]
          example: "published"
        rating:
          type: integer
          minimum: 1
          maximum: 10
          example: 9
        publicReview:
          type: string
          example: "Great place to stay! Very clean and comfortable."
        privateReview:
          type: string
          nullable: true
          example: "Minor issue with WiFi speed"
        reviewCategory:
          type: array
          items:
            $ref: '#/components/schemas/ReviewCategory'
        submittedAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
        guestName:
          type: string
          example: "John Doe"
        listingId:
          type: string
          example: "PROP001"
        listingName:
          type: string
          example: "Modern Downtown Apartment"
        channel:
          type: string
          enum: [Airbnb, Booking.com, Direct, Vrbo, Expedia, Google, Other]
          example: "Airbnb"
        reservationId:
          type: string
          example: "RES123456"
        showOnWebsite:
          type: boolean
          example: true
        responseText:
          type: string
          nullable: true
          example: "Thank you for your feedback!"
        respondedAt:
          type: string
          format: date-time
          nullable: true
          example: "2024-01-16T09:00:00.000Z"
        helpful:
          type: integer
          minimum: 0
          example: 5
        notHelpful:
          type: integer
          minimum: 0
          example: 1
        source:
          type: string
          example: "hostaway"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"

    # Request Types
...
    PaginationParams:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
          example: 10
        total:
          type: integer
          minimum: 0
          example: 150
        pages:
          type: integer
          minimum: 0
          example: 15
    PaginatedReviewsResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Review'
            pagination:
              $ref: '#/components/schemas/PaginationParams'

    # Statistics Types
...
    Property:
      type: object
      properties:
        _id:
          type: string
          example: "60f7b3b3b3b3b3b3b3b3b3b3"
        externalId:
          type: string
          example: "PROP001"
        name:
          type: string
          example: "Modern Downtown Apartment"
        address:
          type: string
          example: "123 Main St, Downtown, City 12345"
        type:
          type: string
          enum: [Apartment, Studio, House, Penthouse, Villa, Other]
          example: "Apartment"
        bedrooms:
          type: integer
          minimum: 0
          maximum: 20
          example: 2
        bathrooms:
          type: integer
          minimum: 0
          maximum: 20
          example: 2
        maxGuests:
          type: integer
          minimum: 1
          maximum: 50
          example: 4
        imageUrl:
          type: string
          nullable: true
          example: "https://example.com/property-image.jpg"
        description:
          type: string
          nullable: true
          example: "Beautiful modern apartment in the heart of downtown"
        amenities:
          type: array
          items:
            type: string
          example: ["WiFi", "Kitchen", "Parking", "Air Conditioning"]
        avgRating:
          type: number
          format: float
          nullable: true
          example: 8.5
        totalReviews:
          type: integer
          nullable: true
          example: 25
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
...
    PropertiesSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Property'
...
    HostawayReviewsResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success]
          example: "success"
        result:
          type: array
          items:
            $ref: '#/components/schemas/Review'
        count:
          type: integer
          example: 25
        source:
          type: string
          enum: [hostaway, mock]
          example: "mock"
...
    GoogleReview:
      type: object
      properties:
        externalId:
          type: string
          example: "google_ChIJN1t_tDeuEmsRUsoyG83frY4_0"
        type:
          type: string
          enum: [guest-to-host]
          example: "guest-to-host"
        status:
          type: string
          enum: [published]
          example: "published"
        rating:
          type: integer
          minimum: 1
          maximum: 10
          example: 8
        publicReview:
          type: string
          example: "Great location and very clean apartment."
        submittedAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
        guestName:
          type: string
          example: "John Smith"
        channel:
          type: string
          enum: [Google]
          example: "Google"
        source:
          type: string
          example: "google"
        profilePhotoUrl:
          type: string
          nullable: true
          example: "https://lh3.googleusercontent.com/a/default-user=s128-c0x00000000-cc-rp-mo"
        relativeTimeDescription:
          type: string
          example: "2 months ago"
    GooglePlaceData:
      type: object
      properties:
        placeName:
          type: string
          example: "FlexLiving Property"
        placeRating:
          type: number
          format: float
          example: 4.5
        totalRatings:
          type: integer
          example: 127
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/GoogleReview'
    GoogleReviewsResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success, error]
          example: "success"
        data:
          $ref: '#/components/schemas/GooglePlaceData'
        message:
          type: string
          example: "Google API key not configured. Returning mock data for demonstration."
        note:
          type: string
          example: "To enable real Google Reviews, add your API key to .env file"
...